Separate out program.c from vm.c
